refactor(fhir): move ResourceFragment data class to core module
